Move home components into a new home folder
