+++ src/components/home/Hero.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../style";
+import styles from "../../style";
 import "../index.css";
 // import { secondimage,thirdimage } from "../assets";
