+++ src/components/home/Outdoor.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
-import styles from "../style";
-import { outDoor } from "../constants";
+import styles from "../../style";
+import { outDoor } from "../../constants";
 import "../index.css";
 import { FaStar } from "react-icons/fa";
